Split unity android extension classes into files
