+++ bugsnag-android-unity/src/main/java/com/bugsnag/android/unity/UnityCallback.java
@@ -1,0 +1,37 @@
+package com.bugsnag.android.unity;
+
+import java.lang.String;
+
+import com.bugsnag.android.*;
+
+class UnityCallback implements Callback {
+    static final String NOTIFIER_NAME = "Bugsnag Unity (Android)";
+    static final String NOTIFIER_VERSION = "3.4.0";
+    static final String NOTIFIER_URL = "https://github.com/bugsnag/bugsnag-unity";
+
+    private final Severity severity;
+    private final String context;
+    private final String logLevel;
+
+    UnityCallback(String context, Severity severity, String logLevel) {
+        this.context = context;
+        this.severity = severity;
+        this.logLevel = logLevel;
+    }
+
+    @Override
+    public void beforeNotify(Report report) {
+        report.setNotifierName(NOTIFIER_NAME);
+        report.setNotifierURL(NOTIFIER_URL);
+        report.setNotifierVersion(NOTIFIER_VERSION);
+
+        com.bugsnag.android.Error error = report.getError();
+        error.setSeverity(severity);
+        MetaData metadata = error.getMetaData();
+        metadata.addToTab("Unity", "unityException", "true");
+        if (logLevel != null && logLevel.length() > 0)
+            metadata.addToTab("Unity", "unityLogLevel", logLevel);
+        if (context != null && context.length() > 0)
+            error.setContext(context);
+    }
+}
